Add third item to stack frame: vars
This separates "env" from "vars" where and environment is a place for functions and global variables, and vars is a place for scoped variables and function parameters. Otherwise defining a function by setting a variable would make the function go out of scope. Or setting loval variables would live longer than a block.