Large and fairly messy refactor of "evalStep"
This introduces a new concept: a frame. This holds local variables, a local
expression stack and a values stack. In order to manage frames in a transparent
way the heap is initialized with a structure that looks like this:

    heap root: (
        (frame <current call frame>)
        (frames ())
        (symbols (
            (MAIN (...))
        ))
    )

The "frame" represents the current frame the code will be executed in. The
"frames" list is a stack of frames. When a `newFrame` is created the current
frame is pushed onto this stack. The "symbols" slot is where all package
symbols are stored, and "MAIN" is where built-in lambda expressions and global
constants are stored - the MAIN symbol list is always resolved when converting
a symbol into something else.

The idea of a frame is to have an easy way to throw away queued expressions,
local variables and local values when a lambda (or block) exits. A frame looks
like this:

    frame: (
        (stack ())
        (variables ())
        (values ())
    )

The "stack" is a list of queued expressions, "variables" are locally defined
variables and "values" are to store the intermediate results of evaluation.