Fix bug in pairSet()
If the key of a pair had a value pairSet() would set the CAR of that value,
rather than swap out the whole cell. I did this so setq() could be used.
Without this setq() would somehow need to be aware of the previous cell, eg:
"setq(car(pair), value)". Here we would need to have access to the ID of "pair"
so we could set it's CDR to a new value. It was due to my misunderstanding of
setq - in reality the "location" in setq is a variable name, so actually setq
should end up calling pairSet() rather than the other way around.