Replace temp hack of "values" reference
The previous code relied on the previous frames "values" reference when
assigning the result of a built-in function.

This new code pushes the value of a built-in function onto the new stack frame
for that function. It then pushes a new special form called "pop-frame" which
pops the latest value from the current active frame and pushes it onto the
previous frame, and then restores that previous frame to the current one.