Change the frame list slightly
Instead of having a current "frame" slot and a "frames" stack, each frame has a
link to it's parent frame. This is effectively the same thing as it makes a
frame a linked list. However it beter expresses something importent: a function
return is taking the latest value from one frame and copying it to its parent
values list.

I did this after thinking a lot about async function calls and channels. In
this expression:

    (* (+ 1 1) (+ 2 2))

in theory the two addition sub-expressions could be executed in parallel. While
there are a host of questions before these two functions could be executed in
parallel, one is: which call frame does a sub-frame return its value to? By
passing the destination (parent) frame when a new frame is created this becomes
explicit. The given frame contains a link to the frame it should return its
value to.

Further issues would be: if the functions take different times to execute, how
to ensure their values are pushed onto the parent stack in the correct order?