Implement "progn"
The "progn" form takes a list of expressions (assumed to be the body of a
function), pushes a "pop-frame" and then the list of expressions in reverse
order. This means each expression will be evaluated in the order they appeared
in the body. The final "pop-frame" takes the last value in the frame and
"returns" it to the previous frame.