Add refcount for objects. Also rename main to builtins
Adding a refcount table for objects. This is basically because of a realization that very often a new cons cell must be created as a copy of an existing cell in order for some basic stuff to work. For example:

    (setq tmp (1 2 3))
    (push list (car tmp))

Pushing an item to a list means setting it's cdr to the rest of the existing
`list`. This is a destructive operation on the `tmp` list because the cdr of
`1` is `2`. So calling `car` on a cons needs to return a copy cell that can
have its cdr freely set without affecting the cdr of the original `1`. However
the cell should point to the same underlying object (the string "1"), hence
needing to track how many cells are pointing to a given object.

The second part of the commit is to move the "main" KV builtin list to a root
key called "builtins". This is because builtins are different than symbols in
the "main" namespace.