Add a "defun" function
This is a bit of an experiment. The "defun" could be implemented as a keyword
(macro) avoiding the need for "quote", but I am still trying to figure out
exactly how macros should work. In particular how to define macros from within
the interpreter, rather than having to write them in Java as special forms.
This implements it as a builtin function, and it does the following:

* Takes three arguments, which are evaluated hence the need for quote - "name",
  "args" and "body".
* Gets the current "syms" entry in the frame and creates a new pair as a
  "lambda".

This is why I rewrote the builtins to receive the whole frame as an argument.

Note: with this in place it is now possible in theory to "bind" a different
"symbols" entry in a frame. This allows for name-spaced packages.