Define functions the "scheme" way
This commit contrasts the previous as a way to define functions. The previous
uses a function (that should really be a macro). This way looks more "schemey"
in that a function definition is simply defining a symbol with a value. To
contrast (removing the "quote"s for clarity):

    (defun sum (a b) (+ 2 (+ a b)))

Vs the "scheme way":

    (define sum (lambda (a b) (+ 2 (+ a b))))

So far I quite like the scheme way :-)