Misc bug fixes, and attempt 1 at macros
* pairSet and pairGet were happily accepting cons 0 as a list. This is bad as
  cons 0 is the special nill cons.
* Deleted "setq" as it made no sense anymore.
* Put a check on "atomString" that is was indeed being called on an atom.
* Removed the usage of "append" in builtin construction. It's still used in Parser...
* In "newFrame" the parent "symbols" was returing zero and this showed up the
  pairSet and pairGet bug. If the parent has no symbols then create a new one.

I also tried an experiment with macros, which is to have a global macros table
that is looked up on each pass. I just push the body onto the stack for eval
with is probably not ideal or even correct. For example it probably needs to be
wrapped in a "progn" so the body is evaluated in the correct order. In addition
to that there is currently no in-language way to access the frame stack,
variables, symbols etc, meaning the macros would be useless basically. I might
need special forms for accessing these from in-lang code.