Experiment with macros
I am having a difficult time figuring out exactly what a macro is. It is
definitely a function because its body must be evaluated. It's very much like a
lambda in that it accepts bound arguments. This experiment attempts to make
lambdas and macros one and the same by *evaluating* the return value from a
function. This means a macro is exactly the same as a function except it
returns an unquoted form instead of a quoted form. The form goes to the stack
where it is evaluated like regular forms.

This does mean however that any function that returns a non-atom form must
return it quoted unless it wants to behave like a macro.