Experiment in macros
I am trying to implement a way of writing custom forms. For example, how could the "while" form be implemented from in-lang (rather than in-java). I came up with a forth-like set of primitives for maniplating the stacks. This is what "while" would look like:

	# expression is pushed onto values stack
	(progn (
	  (.dupv)     // duplicate top of values (the original expression)
	  (.pushs)    // pop value (original expression) onto stack
	  (.cdr)      // pop value from value stack and push cdr onto values (the test)
	  (.dupv)     // dupe the test
	  (.cdr)      // pop the test and push the cdr (body)
	  (.sym then) // create a new symbol and push onto values stack
	  (.list2)    // pop 2 values from value stack and make a new list
	  (.pushs)    // push the new list onto the stack
	  (.pushs)    // push test onto the stack
	)

Note: the "progn" is there to reverse all these operations. For example:

	(while (> x 0) ((let x (- x 1))))

Testing still required.