Playing with Java FFI
I want to be able to load Java classes and use them in-lang seamlessly. This
commit illustrates part of that - attempting to do this with builtin functions.
Previously builtin lambda bodies were a string and in `dispatch()` I just did
string comparison to dispatch to the appropriate function. Here the lambda body
is stored as a `Method` object value, meaning I can just `invoke()` the method
without doing a string lookup.

This will not work however if there are two methods of the same name with
different function signatures. This is because `getMethod()` (or
`getDeclaredMethod()`) requires a list of argument types in order to
pattern-match the correct method. In my case the types are only known at
call-time, meaning I have to look up the Java method then and can not store the
method in the lambda body.

For speed of executing builtins I will probably keep this implementation. In
fact I might drop the whole `variables` construction and let the builtins pop
straight from the values stack - since they do not need local `variables`
bindings as these happen in Java-land. Then I can do the same for FFI Java
calls and pop the values, grabbing their types and building the signature for
`getMethod()`. This will mean that I need some way of casting atoms before
calling FFI Java methods. For example if a Java method requires a Float type I
need to search for that method knowing I am looking for a Float argument
method. But my float might be stored as a String as a regular symbol.

For example, imagine I want to call a Java method with a single Float argument:

* The method is called "floatyBoat"
* There is another method in the same class called "floatyBoat" that accepts a
  String type
* I have a symbol called "myValue" that is a String in the object heap "0.45"

    (floatyBoat myValue)

I look up "floatyBoat" in my symbol table and find it's an FFI Java call. I now
need to locate the Java method called "floatyBoat" but `getMethod()` needs to
be passed either `String.class` or `Float.class` depending on which method I am
trying to get. The string "0.45" is on the values stack but I don't know if I
should call the String method or the Float method. So instead:

    (floatyBoat (J:Float myValue))

Here the `J:Float` is simply a function that pops a String atom from the values
stack, parses it into a Float and pushes that back onto the values stack as a
atomObject with a Float value.

I am note sure if it's a good idea to leave the burden on the programmer. The
only alternative I can think of is to allow multiple symbols with the same name
in the symbols table, add type declarations to the lambda definition somehow.
But then I am making my language typed which I do not want...