Experiments in native Java integration
The plan is to have `jnew` for constructing objects and `jmethod` for calling
methods on those objects. Java objects are stored in the normal object heap.
There is a *lot* of error checking missing here. I introduced a new form called
"HALT" that is an instruction that is "sticky" on the stack (blocking other
instructions) and returns false to indicate something needs to be done. The
Java exceptions return HALT but other things can also return HALT now, like the
Float parsing in plus_e for example.

There is a problem with passing native types (int, float etc) into Java method
calls. This is because I can not create Java objects for these types.