Added "bind" function
This function is to allow switching contexts between namespaces (packages). The
idea is that when you import a function from another package into your current
package, the imported function is wrapped in a "bind". When it is evaluated the
new frame created will be bound to this symbol table instead, meaning local
function calls inside the function refer to the package you are importing the
function from.

Note: there is a bug in that the bind is executed *before* the lambda, and
therefore binds the wrong frame :-/

Bind evaluates the argument (expected to resolve to a symbol name) and looks up
that symbol name in the global sybols list. If not found it creates a new empty
symbol table there under that name. It then sets this symbol table in the
"symbols" slot of the current frame.