This removes the "pop-frame" form
Prior to this, after every lambda body was a pop-frame. This moved the last
item on the values stack of the frame to the values stack of the previous
frame, and then restored the previous frame. This operation is now done by
detecting an empty stack in the current frame. An empty stack means there are
no more instructions to perform, thus an implicit "return".

This is part of the larger picture of moving away from a lambda generating a
new frame. Instead a frame will be created by a "frame" function (or form).
This will allow the programmer to decide the environment that a lambda is
evaluated in. For example importing a function from another namespace should
generate a frame where the "symbols" are referencing the symbol table where the
function is being imported from. Creating a closure should be creating a frame
where "variables" are copied from where the closure was created.

Unfortunately this means a lambda that is not enclosed in a frame will clobber
variables with the same name as input arguments. Recursive lambda's without a
frame will continuously clobber input arguments, and be broken. One possible
solution is to separate "arguments" out from variables in the frame, and do an
extra lookup when resolving symbols.