WIP: Still trying to track down memory leaks
I am doing this by:

1) Descend the root node and flag every cell in use
2) Scan the whole heap looking for used cells

Since everything should be attached to the root node these two numbers should
be the same. They are not - flagged cells are less than used cells. Dumping the
unflagged cells show they are the "builtins" table. For some reason the builtins
table is not being flagged when descending from the root node :-/

It looks like an issue simply with flagging (the cells should be in use), but I
need flagging to work for tracking down future memory leaks.