Failed attempt to use "scope" form function calling
The idea was to make a new function called "call" that would return an
un-quoted "scope" expression (un-quoted expressions returned from functions are
evaluated). The scope expression would set up a new scope with empty variables
and the previous scope's symbol table. Thus acting like a normal function call.

I could then remove the newScope() call in the lambda eval() block. This would
enable the use of lambda without a new scope - for example in a map() function.
It would also allow importing symbols from other namespaces to use the "scope"
function to bind to the symbol table in the namespace.

The problem is the "call" function needs to read the lambda argument list and
pop values from the current scope values to push them onto the new scope.
Otherwise the lambda has no values to read from in the new scope.

I started all this because I realised that a "bind" function to switch the
namespace would operate without a new scope, meaning when done there would be
no way to know how to "pop back" to the previous symbol table.