Get "call" working with variables
This is one way to get "call" working with "scope". The "call" function returns a "scope" expression. This has a quoted stack including the parameters popped from the values stack. When "scope" is evaluated a new scope is created and the values from the previous scope are moved from the expression stack to the values stack (because they have been evaluated already). It works but it is shit because if you wanted to pass a non-atom value you would need to double-quote it - due to this second evaluation.

One way around this would be have the "scope" expression define the "values"
stack. It could then copy the values straight over from the previous scope to
the new scope, avoiding evaluation on the expression stack, thus not needing
double quoting.

Note: in this commit the newScope() function is called still in the "lambda"
block of the eval() function. This is not strictly needed anymore due to "call"
generating the "scope" expression.