Attempt at "import" function
It's broken. The reason is related to how lists work at a fundamental level. A
list is a cell where the CAR points to a list of cells joined by a CAR. When
you call the "symbols" function it will create a new empty list in the global
symbols table. It is a cell with no CAR set because there are no items in the
list. When you copy() an empty list it will create a new cell with a CAR also
set to zero. So you end up with two cells with a different address. When you
push onto one it will not push onto the other one.

I am not sure what the solution should be. Yet.