Handle problem referred to in previous commit
The problem in the previous commit was due to copying lists. Pushing to a
copied list had no effect on the previous list. This commit introduces the
"ref" form, which enables the user to wrap a list in another list thereby
allowing "pass by reference". I made the "symbols" entry in the scope a "ref"
but I should also probably do the same for "variables".

Note: this still does not fix "import" because imported functions are wrapped
in a "scope" expression. This expression does not pull items from the values
stack so when the lambda expression is evaluated there is nothing on the values
stack to pull from - because it's a new empty values stack in the new scope.
There are a few options to solve this:

1) Make the values stack global and not part of a scope - problems here involve
   discarding the return value of void context called functions.

2) Make the "scope" expression provide "values" as a third argument and build
   it by popping: "(reverse (values 2))" - where "values" pops the provided
   number off the values stack into a new list, and "reverse" reverses it.