The end of the "scope" road - or is it...?
This commit represents a not-fully-working but enough-to-demonstrate version of
using "scope" to implement scoped function calls. At this point the "lambda"
form in eval() still creates a new scope, but it doesn't need to. Declaring a
function in the symbol table means storing a scope expression containing the
lambda to be executed in that scope. Importing just needs to store that scope
expression in the importing symbol table. However there are some bits that I
really don't like about this:

1) In order for imported functions to work the "symbols" entry in the "scope"
   expression is stored as a reference to the symbol table. This is a cyclic
   reference since the scope expression is stored in the symbol table.
2) Every single "lambda" expression must be stored wrapped in a "scope",
   otherwise (once the newScope call was removed from the lambda form in eval)
   lambda expressions would never get a separate scope.
3) Processing a scope expression is heavy and makes all function calls slower.

Instead I want functions stored in symbol tables to be stored as regular lambda
expressions, and imported functions to be wrapped in something that binds the
correct symbol table. By default lambda expressions would continue to create
their own scope in the lambda form in eval, as they do right now.

However there is one major disadvantage to this that I was trying to avoid
using "scope": a lambda expression will always create it's own scope. This
basically means I can not do closures, and by extentsion I can probably not
have an object system :-/.

So I am going to keep going and see if I can make things better.