More interactivity via "input"
The idea here is to be able to load more onto the stack at any point in
execution. In the next commit I want to count values on the stack and return a
HALT if there are not enough values to execute a function. This way you could
for example evaluate "+" by itself and the interpreter would HALT with missing
arguments (and push the expression back onto the stack). The user could then
add "(2 3)" which would be evaluated onto the values stack, thus allowing the
"+" to be evaluated. The bigger idea here is that you could interactively "fix"
running code without unwinding the entire stack.