A pretty basic interactive way of evaluating
You can now eval a function without its arguments (eg. "+"). The interpreter
will place a HALT on the stack, blocking the eval loop. You can then place a
"(1 1)" on the stack, which will populate the values stack accordingly. When
your ready you can push a ".popstack" to be eval'ed, which will discard the
HALT and allow the evaluation to continue. So a UI could let you experiement
with functions by running them.